Fix for GA/Wolf Kill Message Duping (#112)
* Added DiedGuardingWolf value

As GA, selecting a wolf who then selects you will result in a kill message being sent twice, once during NightCycle and once during the start of DayCycle. Adding DiedGuardingWolf value to be checked during DayCycle to avoid this.

* Fix GA/Wolf double message issue

When a Guardian Angel guards a wolf on the same night the wolf decides to kill the GA, two messages are sent out announcing the GA's death, one at the moment the GA selects the wolf and the next the morning after when deaths are announced. This patch uses a new value in IPlayer.cs, DiedGuardingWolf, which is set during the night cycle when GA guards wolf and then checked during the day cycle to prevent the second, generic wolf kill message from occurring.