Close #104
Personally I've thought about it many times, but never understood why it shouldn't end, and too many people said this should really end. I really believe there was a logic in this, but until we find it out I'm commenting that line out, because that game should end at least in the majority of the cases.