Enable Check for Cultist Hunter again.
There is a (decent) chance of there being a Cultist game without a Cultist Hunter.
The reason is that the list of possible roles returned by GetRolesList in a cultist game is at least 34 entries long, with 11 players.
That list gets checked for there being a Cultist Hunter in there, however the list is then randomized and only the first 11 entries are used.

As such without the additional check, there is a chance of a cult game being without a Cultist Hunter.