Fixed lowercase at the beginning of a sentence (#211)
When gunner shot default, it looked like this:

Suddenly, a shot rings out.  The village turns to see {hunter} standing over {player}'s body, a gun still aimed at the head. a Werewolf 🐺 is dead.

The "a" in "a werewolf" was lowercase, because the "Wolf"-String was called (of course it could be any role). It also cannot just be made uppercase in the langfile, because it is also used in "{player} was {role}" and then it would look like that: "{player} was A werewolf". What I did is quite much the same as #210, I replaced "IsDead" by "Was", so the dead player's name is the first word in the sentence.