Fix typo within user removal message in vets
Usually,  the user who dint unlock vets will be removed and the previous message showed is "user removed, as they have not unlocked veteran". It sounds like the user removed something . Lets make a small change and make it better like "user was removed, as they have not unlocked veteran".