Fixing AppSeer without Seer bug
The balancing of the rolelist first checked if there was seer when there was appseer, then checked if there was cultist hunter when there is cult. This last step would remove a village role to make room for cultist hunter, so seer could be removed. I fixed the bug by just changing the order. It now checks for CH first, then for seer, to make sure the game never starts with AppSeer without Seer.